Sharding began
We have found a way to complete sharding! We now just need to rewrite a large section of the bot, and how events are handled. This shouldn't be *too* hard, but it will be time consuming

This means that we do not need to convert to reactjs, turns out that wouldn't have helped at all.

Until this is complete, production on the rest of the app has been mostly halted, since it would have to eventually be rewritten